ECONOMY-12 – make Sanity preloads available in nav; more help content to Sanity
